1-portfolo: Switched which components use the tab
* Courses use tabulation; projects use accordian
* Began upgrade to newest ersion of react
    * There is a bug with the current release of tabulation;
       might need to implement custom table
